Replace the 'pyporter2' submodule with a 'fake' submodule managed by kiln.
Rather than forcing everyone to do a 'git submodule init' in order to use
this library, we just check in the submodule as a directory.  This works
well as long as the contents of the submodule don't change very often,
which is the case here (pyporter2 hasn't changed in like 5 years).